refactor(pkg): rebase onto develop — update new bench imports + doc paths
Develop's #1029 added bench/blockstore/concurrent{,_test}.go importing the
pre-rename pkg/blockstore/engine path; repoint to pkg/block/engine. Resolve
CLAUDE.md + docs/ARCHITECTURE.md path tokens against develop's wording.
Verified: go build ./... + go vet -tags integration ./... clean.
