+++ bench/blockstore/concurrent.go
@@ -7,7 +7,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/marmos91/dittofs/pkg/blockstore/engine"
+	"github.com/marmos91/dittofs/pkg/block/engine"
 )
 
 // Concurrent (a)–(d) workloads. Unlike the serial RunWorkload family these
